Rename package from bambu-3mf to bambox
- Rename src/bambu_3mf/ -> src/bambox/
- Update all Python imports, CLI entry point, and pyproject.toml
- Update all docs (CLAUDE.md, README, CONTRIBUTING, ROADMAP, CHANGELOG, SECURITY)
- Update CI workflows (mypy path, coverage module)
- Update Rust bridge references (agent.rs project_name, print_job.rs)
- Update towncrier issue_format URL

The GitHub repo rename (estampo/bambu-3mf -> estampo/bambox) must be
done manually in repo settings. GitHub will redirect the old URL.
