+++ SECURITY.md
@@ -11,14 +11,14 @@
 If you discover a security vulnerability, please report it privately:
 
 1. **Do not** open a public GitHub issue
-2. Email the maintainers or use [GitHub's private vulnerability reporting](https://github.com/estampo/bambu-3mf/security/advisories/new)
+2. Email the maintainers or use [GitHub's private vulnerability reporting](https://github.com/estampo/bambox/security/advisories/new)
 3. Include steps to reproduce, impact assessment, and any suggested fix
 
 We will acknowledge reports within 48 hours and aim to release a fix within 7 days for critical issues.
 
 ## Scope
 
-bambu-3mf handles:
+bambox handles:
 - Archive construction (ZIP files with XML metadata)
 - Credential loading for cloud printing
 - Docker container execution for slicing and bridge communication
