+++ src/bambox/gcode_compat.py
@@ -212,7 +212,7 @@ def _build_header_block(total_layers: int, time_secs: int, max_z: str) -> str:
 
     return (
         "; HEADER_BLOCK_START\n"
-        "; generated by bambu-3mf (generic slicer G-code translated for BBL firmware)\n"
+        "; generated by bambox (generic slicer G-code translated for BBL firmware)\n"
         f"; total estimated time: {time_str}\n"
         f"; total layer number: {total_layers}\n"
         "; filament_diameter: 1.75\n"
