+++ src/bambox/gcode_templates/p1s_start.gcode.j2
