+++ src/bambox/input_spec.example.json
