+++ src/bambox/thumbnail.py
